feat: extract metadata for all databases, add host to config
